backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
pymongo==4.8.0
redis==5.0.7
elasticsearch==8.15.0
python-dotenv==1.0.1
werkzeug==3.0.3
